+++ ioctl.go
@@ -56,39 +56,39 @@ const (
 	_IOCSIZE_MASK  = _IOC_SIZEMASK << _IOC_SIZESHIFT
 )
 
-func _IOC(dir, t, nr, size int) int {
-	return (dir << _IOC_DIRSHIFT) | (t << _IOC_TYPESHIFT) |
-		(nr << _IOC_NRSHIFT) | (size << _IOC_SIZESHIFT)
+func _IOC(dir, t, nr, size int) uintptr {
+	return uintptr((dir << _IOC_DIRSHIFT) | (t << _IOC_TYPESHIFT) |
+		(nr << _IOC_NRSHIFT) | (size << _IOC_SIZESHIFT))
 }
 
-func _IO(t, nr int) int {
+func _IO(t, nr int) uintptr {
 	return _IOC(_IOC_NONE, t, nr, 0)
 }
 
-func _IOR(t, nr, size int) int {
+func _IOR(t, nr, size int) uintptr {
 	return _IOC(_IOC_READ, t, nr, size)
 }
 
-func _IOW(t, nr, size int) int {
+func _IOW(t, nr, size int) uintptr {
 	return _IOC(_IOC_WRITE, t, nr, size)
 }
 
-func _IOWR(t, nr, size int) int {
+func _IOWR(t, nr, size int) uintptr {
 	return _IOC(_IOC_READ|_IOC_WRITE, t, nr, size)
 }
 
-func _IOC_DIR(nr int) int {
-	return ((nr) >> _IOC_DIRSHIFT) & _IOC_DIRMASK
+func _IOC_DIR(nr int) uintptr {
+	return uintptr(((nr) >> _IOC_DIRSHIFT) & _IOC_DIRMASK)
 }
 
-func _IOC_TYPE(nr int) int {
-	return ((nr) >> _IOC_TYPESHIFT) & _IOC_TYPEMASK
+func _IOC_TYPE(nr int) uintptr {
+	return uintptr(((nr) >> _IOC_TYPESHIFT) & _IOC_TYPEMASK)
 }
 
-func _IOC_NR(nr int) int {
-	return ((nr) >> _IOC_NRSHIFT) & _IOC_NRMASK
+func _IOC_NR(nr int) uintptr {
+	return uintptr(((nr) >> _IOC_NRSHIFT) & _IOC_NRMASK)
 }
 
-func _IOC_SIZE(nr int) int {
-	return ((nr) >> _IOC_SIZESHIFT) & _IOC_SIZEMASK
+func _IOC_SIZE(nr int) uintptr {
+	return uintptr(((nr) >> _IOC_SIZESHIFT) & _IOC_SIZEMASK)
 }
